requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anthropic==0.31.2
anyio==4.4.0
attrs==23.2.0
bcrypt==4.2.0
blinker==1.8.2
cachetools==5.4.0
captcha==0.6.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
deepgram-sdk==3.4.0
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
extra-streamlit-components==0.1.71
extra-streamlit-components-better-cookie-manager==0.0.10
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.2
idna==3.7
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain-anthropic==0.1.21
langchain-core==0.2.24
langsmith==0.1.93
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==2.0.1
oauthlib==3.2.2
openai==1.37.1
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.2
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
sentry-sdk==2.13.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SpeechRecognition==3.10.4
streamlit==1.36.0
streamlit-authenticator==0.3.3
streamlit-cookies-controller==0.0.4
streamlit-float==0.3.5
streamlit-js-eval==0.1.7
streamlit-searchbox==0.1.13
streamlit_mic_recorder==0.0.8
tenacity==8.5.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
user-agents==2.2.0
watchdog==4.0.1
websockets==12.0
yarl==1.9.4
